+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+##
+
+- Add support for trimming.
+
 ## 0.4.1
 
 - Fix callbacks that listen to derived event types not being raised unless the derived type has at least one subscribed callback.