+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ##
 
 - Add support for trimming.
+- Improve internal array pool sharing (reduces memory consumption and allocation) in .NET >= 5.
 
 ## 0.4.1