+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Add support for trimming.
 - Improve internal array pool sharing (reduces memory consumption and allocation) in .NET >= 5.
+- Remove dependencies in .NET >= 5.
+- Replace `System.Buffers` and `System.Runtime.CompilerServices.Unsafe` for `System.Memory` dependency in .NET < 5.
 
 ## 0.4.1